hints, instructions
