+++ languages/elixir/exercises/concept/agent/lib/community_garden.ex
@@ -1,45 +1,24 @@
 defmodule Plot do
-  @enforce_keys [:plot_id, :registered_to]
-  defstruct [:plot_id, :registered_to]
+  raise "Please implement the Plot struct"
 end
 
 defmodule CommunityGarden do
-  def start(opts \\ []) do
-    {initial_registry, opts} = Keyword.pop(opts, :initial_registry, %{})
-    {initial_next_id, opts} = Keyword.pop(opts, :initial_next_id, 1)
-    Agent.start(fn -> %{registry: initial_registry, next_id: initial_next_id} end, opts)
+  def start(opts) do
+    raise "Please implement the start/1 function"
   end
 
   def list_registrations(pid) do
-    Agent.get(pid, fn state ->
-      state.registry
-    end)
+    raise "Please implement the list_registrations/1 function"
   end
 
   def register(pid, register_to) do
-    Agent.get_and_update(pid, fn %{registry: registry, next_id: next_id} = state ->
-      new_plot = %Plot{plot_id: next_id, registered_to: register_to}
-      updated = Map.put(registry, next_id, new_plot)
-      {new_plot, %{state | registry: updated, next_id: next_id + 1}}
-    end)
+    raise "Please implement the register/2 function"
   end
 
   def release(pid, plot_id) do
-    Agent.update(pid, fn %{registry: registry} = state ->
-      updated = Map.delete(registry, plot_id)
-      %{state | registry: updated}
-    end)
+    raise "Please implement the release/2 function"
   end
 
   def get_registration(pid, plot_id) do
-    registration =
-      Agent.get(pid, fn state ->
-        state.registry[plot_id]
-      end)
-
-    case registration do
-      nil -> {:not_found, "plot is unregistered"}
-      _ -> registration
-    end
-  end
+    raise "Please implement the get_registration/2 function"
 end
